[FIX] 대출/반납 시 연체일 조건 수정 (#16) - 대출 시 연체일이 지났으면 연체일을 초기화 시켜준다. - 반납시 연체일이 지나있으면 현재 시간부터 연체일을 부과합니다.

+++ backend/library/src/main/java/com/hae/library/service/BookService.java
@@ -163,11 +163,19 @@ public class BookService {
         // 삭제할 도서를 ID를 이용하여 검색합니다. 도서를 찾지 못하면 예외를 발생시킵니다.
         Book book = bookRepo.findById(bookId).orElseThrow(() -> new RestApiException(BookErrorCode.BAD_REQUEST_BOOK));
 
-        // TODO: 도서가 대여중인지 확인
-        // TODO: 마지막 도서이면 BookInfo도 삭제
+        // TODO: 도서가 대여중인지 확인합니다.
+        if (book.isLendingStatus() == true) {
+            throw new RestApiException(BookErrorCode.NOT_DELETE_BECAUSE_RENT_BOOK);
+        }
+
+        // TODO: 마지막 도서이면 BookInfo도 삭제합니다.
+        if (bookRepo.countByBookInfo(book.getBookInfo()) == 1) {
+            bookInfoRepo.delete(book.getBookInfo());
+        }
 
         // 찾은 도서를 데이터베이스에서 삭제합니다.
         bookRepo.deleteById(bookId);
+
     }
 
 }
